Enable SO_NOSIGPIPE on macOS and *BSD (#1028)
* Enable SO_NOSIGPIPE on macOS and *BSD

When Z_FEATURE_AUTO_RECONNECT is enabled, zenoh-pico gets terminated
by macOS with signal 13 (SIGPIPE) and thus the reconnection fails
every time.

On Linux, zenoh-pico sets MSG_NOSIGNAL in send but it explicitly
disables SO_NOSIGPIPE on macOS and *BSD. This behavior originates
in the second commit of this repository: 71eb5891c.

* Remove *BSD-specific code in system/windows/network.c