enhancement: zephyr: use cmake config generation
The standalone build chain relies on cmake generating various config
files based off cmake variables and template files. The zephyr cmake
chain did not follow this approach.

Due to this, compiling zenoh-pico as a west module inside a zephyr
project would for example lead to warnings about redefinitions of the
various Z_FEATURE_ macros that zenoh-pico relies on for controlling the
feature set.

Update the zephyr CMakeLists to use the existing feature_config.cmake
helper, which takes care of generating the feature configuration files
based off the various cmake variables. Since certain features are
accessible via. Kconfig options, the corresponding cmake variables are
set prior to inclusion of the feature_config helper, effectively
overriding the default values.

In addition, the zephyr CMakeLists is updated to fetch the zenoh-pico
version from the `version.txt` file in the root of the project. This
version is then forwarded to another configuration file via. cmake. This
mimics the behavior of the CMakeLists from the root of the project.

Since the zephyr cmake chain runs with a different base directory than a
standalone build, the feature_config.cmake is updated to rely on a
ZENOH_PICO_INCLUDEDIR variable for locating the respective configuration
files. The root CMakeLists is updated to set this appropriately.

These changes were based off the draft pull-request with zephyr build
fixes: https://github.com/eclipse-zenoh/zenoh-pico/pull/949

Signed-off-by: Emil Dahl Juhl <emil@s16s.ai>