seqnumber: Fix C++ interopability
_Atomic and std::atomic are not guaranteed to be compatible.
Additionally, std::atomic introduces dependencies on the C++ standard library,
which is often too large for resource-constrained embedded systems.
This change moves all atomic operations into seqnumber.c and uses the C11
atomic_xx operations, which are safe for both C and C++ callers.
It also fixes a bug where _z_seqnumber_null was not implemented as an atomic operation.