seqnumber: Fix C++ interopability and mutex cleanup (#1092)
* seqnumber: Fix C++ interopability

_Atomic and std::atomic are not guaranteed to be compatible.
Additionally, std::atomic introduces dependencies on the C++ standard library,
which is often too large for resource-constrained embedded systems.
This change moves all atomic operations into seqnumber.c and uses the C11
atomic_xx operations, which are safe for both C and C++ callers.
It also fixes a bug where _z_seqnumber_null was not implemented as an atomic operation.

* Ensure sequence number is correctly cleaned up in the case of C99 where a mutex is present.

---------

Co-authored-by: Peter van der Perk <peter.vanderperk@nxp.com>